requirements.txt
mysql-connector-python==8.0.33
pyserial==3.5
